Move api.gui.igredient to a correctly spelled package (#25)
