+++ src/main/java/gregtech/api/gui/ingredient/IIngredientSlot.java
@@ -1,4 +1,4 @@
-package gregtech.api.gui.igredient;
+package gregtech.api.gui.ingredient;
 
 public interface IIngredientSlot {
