+++ src/main/java/gregtech/api/gui/widgets/AbstractWidgetGroup.java
@@ -3,8 +3,8 @@ package gregtech.api.gui.widgets;
 import gregtech.api.gui.INativeWidget;
 import gregtech.api.gui.IRenderContext;
 import gregtech.api.gui.Widget;
-import gregtech.api.gui.igredient.IGhostIngredientTarget;
-import gregtech.api.gui.igredient.IIngredientSlot;
+import gregtech.api.gui.ingredient.IGhostIngredientTarget;
+import gregtech.api.gui.ingredient.IIngredientSlot;
 import gregtech.api.util.Position;
 import gregtech.api.util.Size;
 import mezz.jei.api.gui.IGhostIngredientHandler.Target;
