+++ src/main/java/gregtech/api/gui/widgets/PhantomFluidWidget.java
@@ -4,8 +4,8 @@ import com.google.common.collect.Lists;
 import gregtech.api.gui.GuiTextures;
 import gregtech.api.gui.IRenderContext;
 import gregtech.api.gui.Widget;
-import gregtech.api.gui.igredient.IGhostIngredientTarget;
-import gregtech.api.gui.igredient.IIngredientSlot;
+import gregtech.api.gui.ingredient.IGhostIngredientTarget;
+import gregtech.api.gui.ingredient.IIngredientSlot;
 import gregtech.api.gui.resources.RenderUtil;
 import gregtech.api.gui.resources.TextureArea;
 import gregtech.api.util.Position;
