+++ src/main/java/gregtech/common/EventHandlers.java
@@ -1,6 +1,7 @@
 package gregtech.common;
 
 import gregtech.api.GTValues;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.monster.EntityEnderman;
 import net.minecraft.init.Items;
 import net.minecraft.init.MobEffects;
@@ -8,9 +9,11 @@ import net.minecraft.item.ItemStack;
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.ConfigManager;
 import net.minecraftforge.event.entity.living.EnderTeleportEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.fml.client.event.ConfigChangedEvent;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 @Mod.EventBusSubscriber
@@ -45,5 +48,12 @@ public class EventHandlers {
         if (event.getModID().equals(GTValues.MODID)) {
             ConfigManager.sync(GTValues.MODID, Config.Type.INSTANCE);
         }
+    }
+
+    @SubscribeEvent(priority = EventPriority.HIGHEST)
+    public void onBreakSpeed(PlayerEvent.BreakSpeed event)
+    {
+        IBlockState state = event.getState();
+
     }
 }
